git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/FgpUtil/trunk@12393 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Util/src/test/java/org/gusdb/fgputil/db/runner/SQLRunnerQueryTest.java
@@ -48,7 +48,7 @@ public class SQLRunnerQueryTest {
     
     assertEquals(_handler.getNumRows(), 2);
     assertEquals(_handler.getNumCols(), 3);
-    assertEquals(((Integer)_handler.getColumnTypes().get(2)).intValue(), Types.VARCHAR);
+    assertEquals(_handler.getColumnTypes().get(2).intValue(), Types.VARCHAR);
     
     /* Debug printouts
     printRow(_handler.getColumnNames());
